feat: add streamlined version of beta debiasing logic
this ports in the debiasing logic from the original implementation. only logic used for linear
regression is added here, as that was the only bit that was actually supported originally, and the
rest represented some efforts to extend to other GLMs that didn't culminate in much. those features
may be added back in at a later date