chore: add streamlined standard error expectation for linear regression
during the original development process of this package, I made various attempts at adjusting
standard error to reflect the debiasing process. unfortunately, these never worked very well, and as
mentioned in various other places, non-linear GLM support never really worked. I'm leaving this bit
in here to record the closed form expectation of standard error for an additive genetic component,
as it may be useful at some point in the future and was not captured in the original code