feat: add primary logic for debiasing and MSE correction of linear regression data
this ports in the remaining primary logic from the original implementation. several alternate run
modes are excluded, as they had been initially intended for debugging; they may be added back. this
reimplementation once again removes a large amount of logic for GLM support outside of linear
regression. this also revealed some hackjob nonsense that I had to add at the end to override MSE
correction, as it never seemed to work right. this will need to be addressed now.